Finished LogWriter and corresponding unit test

+++ DataChecker/DataCheckerTest/ErrorHandling/LogWriter.cs
@@ -8,6 +8,7 @@ namespace DataCheckerTest.ErrorHandling
     [TestClass]
     public class LogWriterClassTest
     {
+        private bool deleteLogFile = true; // just in case log file already exists in file system and shouldnt be deleted
         private string logFile;
 
         [TestMethod]
@@ -66,12 +67,70 @@ namespace DataCheckerTest.ErrorHandling
             File.Delete(logFile); // DELETE file after test, cleanup
         }
 
+        [TestMethod]
+        public void TestLogInformation()
+        {
+            /* Method doesn't test Timestamp/DateTime.Now functionality 
+             * because desired approach to making timestamps testable (provider injection)
+             * required dependency inaccessible on deployment environment...resulting in need to change code before deployment
+             * https://docs.microsoft.com/en-gb/archive/blogs/ploeh/provider-injection
+             * */
+            string logFolder = Environment.CurrentDirectory + @"\";
+            logFile = logFolder + "table.txt";
+            string firstExpectedOutput = "|| INFO|| source || table || info";
+            string secondExpectedOutput = "|| INFO|| source2 || table || info2";
+
+            if (File.Exists(logFile))
+            {
+                Assert.IsTrue(false, "file used for testing " + logFile + " already exists. Cancelling test.");
+                return;
+            }
+
+            LogWriter.LogInformation("source", "table", "info", logFolder);
+
+            if (!File.Exists(logFile))
+            {
+                Assert.IsTrue(false, "log file was not created as expected.");
+            }
+            else
+            {
+                string[] readText = File.ReadAllLines(logFile);
+                string actualOutputWithoutTimestamp = readText[0].Substring(readText[0].IndexOf("||"), readText[0].Length - readText[0].IndexOf("||")); // trim timestamp because untested
+
+                Assert.AreEqual(1, readText.Length); // expect 1 line of text
+                Assert.AreEqual(firstExpectedOutput, actualOutputWithoutTimestamp); // expect first line of file to match expectation
+            }
+
+            /* test successive writes to same file */
+            LogWriter.LogInformation("source2", "table", "info2", logFolder);
+
+            if (!File.Exists(logFile))
+            {
+                Assert.IsTrue(false); // fail test because expect this file to be created
+            }
+            else
+            {
+                string[] readText = File.ReadAllLines(logFile);
+                string actualOutputWithoutTimestamp = readText[0].Substring(readText[0].IndexOf("||"), readText[0].Length - readText[0].IndexOf("||")); // trim timestamp because untested
+                string actualOutputWithoutTimestamp2 = readText[1].Substring(readText[1].IndexOf("||"), readText[1].Length - readText[1].IndexOf("||")); // trim timestamp because untested
+
+                Assert.AreEqual(2, readText.Length); // expect 2 line of text and 1 new line
+                Assert.AreEqual(firstExpectedOutput, actualOutputWithoutTimestamp); // expect first line of file to match expectation
+                Assert.AreEqual(secondExpectedOutput, actualOutputWithoutTimestamp2); // expect second line of file to match expectation
+            }
+
+            File.Delete(logFile); // DELETE file after test, cleanup
+        }
+
         [TestCleanup]
         public void PostExecution()
         {
-            if (File.Exists(logFile))
+            if (deleteLogFile)
             {
-                File.Delete(logFile);
+                if (File.Exists(logFile))
+                {
+                    File.Delete(logFile);
+                }
             }
         }
     }
